Changes required for using CCXT as the broker

+++ sysexecution/tick_data.py
@@ -176,7 +176,7 @@ class tickerObject(object):
     We wrap it in this so have standard methods
     """
 
-    def __init__(self, ticker, qty: int = arg_not_supplied):
+    def __init__(self, ticker, qty: float = arg_not_supplied):
         # 'ticker' will depend on the implementation
         self._ticker = ticker
         self._qty = qty
@@ -187,7 +187,7 @@ class tickerObject(object):
         return self._ticker
 
     @property
-    def qty(self) -> int:
+    def qty(self) -> float:
         qty = self._qty
         return qty
 
@@ -241,7 +241,7 @@ class tickerObject(object):
     def analyse_for_tick(
         self,
         tick: oneTick = arg_not_supplied,
-        qty: int = arg_not_supplied,
+        qty: float = arg_not_supplied,
         replace_qty_nans=False,
     ):
         if qty is arg_not_supplied:
@@ -258,7 +258,7 @@ class tickerObject(object):
         return results
 
     def wait_for_valid_bid_and_ask_and_analyse_current_tick(
-        self, qty: int = arg_not_supplied, wait_time_seconds: int = 10
+        self, qty: float = arg_not_supplied, wait_time_seconds: int = 10
     ) -> oneTick:
 
         current_tick = self.wait_for_valid_bid_and_ask_and_return_current_tick(
